Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -4,7 +4,7 @@
 int main() {
     bool result=false;
     long a;
-    int b=pow(2,31);
+    long b=pow(2,31);
     scanf("%ld",&a);
     result= a>=b;
     if (result){
